// shared/src/gen/type-move-meta.ts
export const MoveTypeMap = {
// ...
   "Bide": { name: "Bide", type: 'Normal' as Type, attackKey: 'attack', power: 0, accuracy: Infinity, flavor: "The user accepts damage for several turns before unleashing a strong attack." },
// ...
   "Dig": { name: "Dig", type: 'Ground' as Type, attackKey: 'attack', power: 1.1, accuracy: 1, flavor: "The user burrows into the ground and strikes the target when they least expect it." },
// ...
   "Dive": { name: "Dive", type: 'Water' as Type, attackKey: 'attack', power: 1.1, accuracy: 1, flavor: "The user hides underwater and suddenly rises and sneak attacks the target." },
// ...
   "Fly": { name: "Fly", type: 'Flying' as Type, attackKey: 'attack', power: 1.1, accuracy: 0.95, flavor: "The user flies into the air and later strikes the target." },
// ...
   "Geomancy": { name: "Geomancy", type: 'Fairy' as Type, attackKey: 'attack', power: 0, accuracy: Infinity, flavor: "The user draws the power of life itself to boost its stats over two turns." },
// ...
   "Outrage": { name: "Outrage", type: 'Dragon' as Type, attackKey: 'attack', power: 1.4, accuracy: 1, flavor: "The user might start a tantrum that causes serious damage. Then it becomes confused." },
// ...
   "Struggle": { name: "Struggle", type: 'Status' as Type, attackKey: 'attack', power: 0.5, accuracy: Infinity, flavor: "The user fights with all it can do. It hurts itself too." },
// ...
   "Thrash": { name: "Thrash", type: 'Normal' as Type, attackKey: 'attack', power: 1.4, accuracy: 1, flavor: "The user goes on an angry rampage. They become confused later on." },
// ...
}
// ...
export type MoveId = keyof typeof MoveTypeMap